==> Final_Fig4/step1.1_scatterplot between biological replicates.Rmd <==
---
title: "STEP1.1:scatterplot between biological replicates"
author: "Yeonui+Kwak"
date: "8/13/2020"
output: html_document
---




scatterplot to see correlation between two biological replicates.
ted-seq scatterplot for all timepoints.
```{R}

# Correlation scatterplots of PRO-seq data
library(dplyr)
library(tidyr)
library(ggplot2)
library(gridExtra)
source("scatterPlot.R")
```
3.Load the data
```{R}
pal1<-read.table("11868.PASisoforms_meanPAL_rep1.txt",header=T, stringsAsFactors = F)
nrow(pal1)
colnames(pal1)=c("newid","oldid",paste("rep1",c(0,1,2,4),sep="_"))

pal2<-read.table("12258.PASisoforms_meanPAL_rep2.txt",header=T, stringsAsFactors = F)
colnames(pal2)=c("newid","oldid",paste("rep2",c(0,1,2,4),sep="_"))
final=inner_join(pal1,pal2,by="newid")
#final<-read.table("REAL.final_PAL_data.txt",sep="\t",head=T,stringsAsFactors = F)
final%>%drop_na()%>%nrow()#10077 

```

4. Draw a scatterplot between 0h and 1h.
```{R}
getwd()
t=final%>% select(rep1_0,rep2_0)%>%mutate(x=rep1_0,y=rep2_0)
cat("correlation",cor(t$x,t$y))
t=t%>%mutate(density = densCols(log(x)/log(2),log(y)/log(2),
                              nbin=512, bandwidth = 0.1,
                              colramp = colorRampPalette(yk.col(5))))

pdf("0and0:PAL_replicates_scattorplot.pdf", width = 3, height = 3)
ggplot(t, aes(y = y, x = x)) + 
    geom_point(aes(col = density), size=1) + scale_color_identity() +
    theme_bw() + 
    labs(x ="PAL.rep1.0h (nt)", y ="PAL.rep2.0h (nt)")+xlim(c(0,300))+ylim(c(0,300))#+
  #geom_abline(slope = 1, intercept = 20,linetype="dashed")+
  #geom_abline(slope = 1, intercept = -20,linetype="dashed")+
  #geom_abline(slope = 1, intercept = 0,col="grey",size=1, linetype="dotted")+
  #ggtitle("TED-seq: 0h")
dev.off()


t=final%>% select(rep1_1,rep2_1)%>%mutate(x=rep1_1,y=rep2_1)
cat("correlation",cor(t$x,t$y))
t=t%>%mutate(density = densCols(log(x)/log(2),log(y)/log(2),
                              nbin=512, bandwidth = 0.1,
                              colramp = colorRampPalette(yk.col(5))))

pdf("1and1:PAL_replicates_scattorplot.pdf", width = 3, height = 3)
ggplot(t, aes(y = y, x = x)) + 
    geom_point(aes(col = density), size=1) + scale_color_identity() +
    theme_bw() + 
    labs(x ="PAL.rep1.1h (nt)", y ="PAL.rep2.1h (nt)")+xlim(c(0,300))+ylim(c(0,300))#+
  #geom_abline(slope = 1, intercept = 20,linetype="dashed")+
  #geom_abline(slope = 1, intercept = -20,linetype="dashed")+
  #geom_abline(slope = 1, intercept = 0,col="grey",size=1, linetype="dotted")+
  #ggtitle("TED-seq: 0h")
dev.off()

t=final%>% select(rep1_2,rep2_2)%>%mutate(x=rep1_2,y=rep2_2)
cat("correlation",cor(t$x,t$y))
t=t%>%mutate(density = densCols(log(x)/log(2),log(y)/log(2),
                              nbin=512, bandwidth = 0.1,
                              colramp = colorRampPalette(yk.col(5))))

pdf("2and2:PAL_replicates_scattorplot.pdf", width = 3, height = 3)
ggplot(t, aes(y = y, x = x)) + 
    geom_point(aes(col = density), size=1) + scale_color_identity() +
    theme_bw() + 
    labs(x ="PAL.rep1.2h (nt)", y ="PAL.rep2.2h (nt)")+xlim(c(0,300))+ylim(c(0,300))#+
  #geom_abline(slope = 1, intercept = 20,linetype="dashed")+
  #geom_abline(slope = 1, intercept = -20,linetype="dashed")+
  #geom_abline(slope = 1, intercept = 0,col="grey",size=1, linetype="dotted")+
  #ggtitle("TED-seq: 0h")
dev.off()

t=final%>% select(rep1_4,rep2_4)%>%mutate(x=rep1_4,y=rep2_4)
cat("correlation",cor(t$x,t$y))
t=t%>%mutate(density = densCols(log(x)/log(2),log(y)/log(2),
                              nbin=512, bandwidth = 0.1,
                              colramp = colorRampPalette(yk.col(5))))

pdf("4and4:PAL_replicates_scattorplot.pdf", width = 3, height = 3)
ggplot(t, aes(y = y, x = x)) + 
    geom_point(aes(col = density), size=1) + scale_color_identity() +
    theme_bw() + 
    labs(x ="PAL.rep1.4h (nt)", y ="PAL.rep2.4h (nt)")+xlim(c(0,300))+ylim(c(0,300))#+
  #geom_abline(slope = 1, intercept = 20,linetype="dashed")+
  #geom_abline(slope = 1, intercept = -20,linetype="dashed")+
  #geom_abline(slope = 1, intercept = 0,col="grey",size=1, linetype="dotted")+
  #ggtitle("TED-seq: 0h")
dev.off()
```
