==> Final_Fig4/step4_PAL dynamics.Rmd <==
---
title: "Step4: PAL temporal dynamics"
author: "Yeonui+Kwak"
date: "8/13/2020"
output: html_document
---

1.Set a working directory
```{R}
setwd("/Users/yeonuikwak/Desktop/Publication_Mar/Final_Fig4/Rcode/step4")

```
2.Load the required packages,and set up colorcode.
```{r}
library(RColorBrewer)
#nstall.packages("raster")
library(raster)
library(tidyverse)
library(dplyr)
library(tidyr)
library(ggplot2)
library(ggbeeswarm)

#BiocManager::install("ComplexHeatmap")
library(ComplexHeatmap)
### Heatmap with Annotation
library(dendextend)
library("circlize")
library("RColorBrewer")
library(cluster)    # clustering algorithms
library(factoextra)


yk.col = function(n = 5, sat = 0.7, lum = 1) {
	col = c("royalblue4", "deepskyblue3", "turquoise3",
	   "grey62", "goldenrod2", "orange2", "orangered2") %>%
	col2rgb %>%
	rgb2hsv
	return(colorRampPalette(hsv(col[1,], col[2,] * sat,
				    col[3,] * lum))(n)) }
```

3.Load the PAL file.
```{R}
custom<-read.table("custom.meanPAL.acrossalltimepoints.txt",header=T,stringsAsFactors = F)
id<-read.csv("PALchange.transcriptid.list.txt",sep="\t",stringsAsFactors = F)
final=custom%>%filter(newid %in% unlist(id$newid))
final%>%arrange(hgnc)%>%select(hgnc)%>%distinct()%>%nrow() #1404
final%>%arrange(newid)%>%select(newid)%>%distinct()%>%nrow() #1520
head(final)
```

```{R}
write.table(final%>%unite("PASid",c(hgnc,ALE,tandem.new),sep=":",remove=F),"PALdynamics.n.1520.transcriptisoforms.txt"
,row.names = F,col.names = T,sep="\t",quote=F)
final=final%>%unite("PASid",c(hgnc,ALE,tandem.new),sep=":",remove=F)%>%select(PASid,hgnc,X0,X1,X2,X4)
final[,3:6]=(final[,3:6]-apply(final[,3:6],1,mean))
#custom[,2:5]=(custom[,2:5]-custom[,2])
colnames(final)=c("PASid","hgnc","0h","1h","2h","4h")
```



```{R}
#col<-RColorBrewer::brewer.pal(8,"Paired")
k2 <- kmeans(final[,3:6], centers=9  , nstart = 25)

highly_Var = as.tibble(final)%>%
  mutate("cl"=k2$cluster)%>%
  group_by(cl)%>%arrange(cl,desc(`0h`),desc(`1h`),desc(`2h`),desc(`4h`))%>%
  ungroup()%>%mutate(cl=factor(cl,levels=c(1:9)))%>%mutate(newcl=ifelse(cl==3,4,
                                                                        ifelse(cl==4,3,
                                                                               ifelse(cl==5,8,
                                                                                      ifelse(cl==7,5,
                                                                                             ifelse(cl==1,9,
                                                                                                    ifelse(cl==2,2,                                                                                                           ifelse(cl==6,6,
                                                                                                                                                                                                                                     ifelse(cl==8,1,ifelse(cl==9,7,cl))))))))))
#TNF
highly_Var%>%filter(grepl("TNF",hgnc))
highly_Var%>%filter(grepl("ENST00000449264",newid))
nrow(highly_Var)
write.table(highly_Var,"9cluster.txt",sep="\t",col.names = T,row.names = F,quote = F)
```

```{r}
highly_Var<-read.table("9cluster.txt",stringsAsFactors = F,header=T)
colnames(highly_Var)[3:6]=c(0,1,2,4)
ha = HeatmapAnnotation(df = data.frame(TimeuponLPS= as.factor(c(0,1,2,4))))
pdf("Fig4C_PALdynamics_n1520.pdf",height=6,width=4)
Heatmap(highly_Var[,3:6], name = "PAL-meanPAL", 
        show_row_names = F,cluster_columns = F,show_column_names = TRUE,  split=highly_Var$newcl,col= circlize::colorRamp2(seq(-20,20,5), yk.col(9)))
dev.off()


head(highly_Var)
pdf("F4C_PALdynamics_n1520_linecol8.pdf",height=8,width=2.5)
highly_Var%>%gather("time","pal",-PASid,-newcl,-cl,-hgnc)%>%
  mutate(time=as.numeric(time))%>%group_by(PASid)%>%ungroup()%>%ggplot(aes(time,pal,group=PASid))+geom_line(aes(col=as.factor(newcl),alpha=0.2))+facet_grid(newcl~.,scales="fixed",switch="both")+ylab("PAL")+
  scale_color_manual(values =yk.col(9),name = "Cluster" )+theme(strip.text = element_text(size=12))+theme(legend.position = "none")+
  theme_bw()
dev.off()
nrow(highly_Var)
```


Enlarge the cluster 5,6,7
```{R}
subset=highly_Var%>%filter(newcl%in%c(4,5,6))
subset=subset%>%arrange(hgnc)%>%group_by(hgnc)%>%mutate(index=row_number())%>%unite("newhgnc",c(hgnc,index),sep=": ")
t=subset[,3:6]
rownames(t)=subset$newhgnc
pdf("cluster4.5.6.pdf",height=15,width=6)
Heatmap(t, name = "PAL-meanPAL", 
        show_row_names = T,cluster_columns = F, show_column_names = TRUE,  split=subset$newcl,col= circlize::colorRamp2(seq(-20,20,5), yk.col(9)))
dev.off()

```


gene list per cluster
```{R}
custom<-read.table("custom.meanPAL.acrossalltimepoints.txt",header=T,stringsAsFactors = F)
id<-read.csv("PALchange.transcriptid.list.txt",sep="\t",stringsAsFactors = F)
final=custom%>%filter(newid %in% unlist(id$newid))
final%>%arrange(hgnc)%>%select(hgnc)%>%distinct()%>%nrow() #1404
final%>%arrange(newid)%>%select(newid)%>%distinct()%>%nrow() #1520
head(final)

final=final%>%unite("PASid",c(hgnc,ALE,tandem.new),sep=":",remove=F)%>%select(PASid,newid,hgnc,X0,X1,X2,X4)
highly_Var.tmp=highly_Var%>%inner_join(final%>%select(PASid,newid),by="PASid")

for (i in 1:9){
  write.table(highly_Var.tmp%>%filter(newcl==i)%>%select(newid)%>%separate(newid,c("tid","no"),sep="\\.")%>%select(tid)%>%arrange(tid)%>%distinct(),paste("paldynamics.cl",i,
                                                             ".txt",sep=""),quote = F,col.names = F,row.names = F)
}

#background

bg=read.table("custom.meanPAL.acrossalltimepoints.txt",header=T,stringsAsFactors = F)
nrow(bg)
write.table(bg%>%select(newid)%>%separate(newid,c("tid","no"),sep="\\.")%>%select(tid)%>%arrange(tid)%>%distinct(),"paldynamics.background.txt",quote = F,col.names = F,row.names = F)
head(bg)




#summary

t=read.table("9cluster.txt",header=T,stringsAsFactors = F)
write.table(t%>%group_by(newcl)%>%summarise(count=n()),"summary_9cluster.txt",col.names=T,row.names = F,sep="\t",quote=F)
```
